[MOVE] Auth / dto에 대해서 request, response 폴더 생성
